Update to read sinalpha and cosalpha and use double precision
Previously sinalpha and cosalpha were recomputed in every image (and this will still happen if they are not provided as input.
However, this could result in artifacts on the image boundaries, by reading sinalpha and cosalpha this artifact is
removed.

+++ src/objects/domain_h.f90
@@ -89,8 +89,8 @@ module domain_interface
     real :: dx
     integer :: nsmooth
 
-    real, allocatable :: costheta(:,:)
-    real, allocatable :: sintheta(:,:)
+    double precision, allocatable :: costheta(:,:)
+    double precision, allocatable :: sintheta(:,:)
 
     ! these coarrays are used to send all data to/from a master image for IO... ?
     ! For now this will be taken care of in the boundary conditions object
